+++ src/main/java/it/univaq/disim/GymREST/resources/FeedbackGymRes.java
@@ -1,14 +1,16 @@
 package it.univaq.disim.GymREST.resources;
 
 import it.univaq.disim.GymREST.business.FeedbackGymService;
+import it.univaq.disim.GymREST.business.UserService;
 import it.univaq.disim.GymREST.business.impl.FeedbackGymServiceImpl;
+import it.univaq.disim.GymREST.business.impl.UserServiceImpl;
 import it.univaq.disim.GymREST.model.FeedbackGym;
+import it.univaq.disim.GymREST.model.User;
+import it.univaq.disim.GymREST.security.Auth;
 
+import javax.annotation.security.RolesAllowed;
 import javax.ws.rs.*;
-import javax.ws.rs.core.Context;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.UriInfo;
+import javax.ws.rs.core.*;
 import java.sql.SQLException;
 
 public class FeedbackGymRes {
@@ -41,37 +43,60 @@ public class FeedbackGymRes {
     }
 
     @POST
+    @Auth
     @Consumes(MediaType.APPLICATION_JSON)
-    public Response addFeedbackGym(@Context UriInfo uriinfo, FeedbackGym feedbackGym) throws SQLException {
-        FeedbackGymService feedbackGymService = new FeedbackGymServiceImpl(urlDB, userDB, pswDB);
-        feedbackGym.setGym(idGym);
+    public Response addFeedbackGym(@Context SecurityContext securityContext, @Context UriInfo uriinfo, FeedbackGym feedbackGym) throws SQLException {
+        String username = securityContext.getUserPrincipal().getName();
+        if (securityContext.isUserInRole("utente")) {
+            UserService userService = new UserServiceImpl(urlDB, userDB, pswDB);
+            User user = userService.getUserByUsername(username);
 
-        //recupera id da utente connesso
-        feedbackGym.setUser(1);
+            FeedbackGymService feedbackGymService = new FeedbackGymServiceImpl(urlDB, userDB, pswDB);
+            feedbackGym.setUser(user.getId());
+            feedbackGym.setGym(idGym);
 
-        long idFeedback = feedbackGymService.createFeedbackGym(feedbackGym);
+            long idFeedback = feedbackGymService.createFeedbackGym(feedbackGym);
 
-        return Response.created(uriinfo.getAbsolutePathBuilder().path(this.getClass(), "getFeedbackGym").build(idFeedback)).build();
+            return Response.created(uriinfo.getAbsolutePathBuilder().path(this.getClass(), "getFeedbackGym").build(idFeedback)).build();
+        } else {
+            return Response.serverError().entity("Non hai i permessi per fare questa operazione").build();
+
+        }
     }
 
     @PUT
+    @Auth
     @Path("{idFeedback: [0-9]+}")
     @Consumes(MediaType.APPLICATION_JSON)
-    public Response updateFeedbackGym(@PathParam("idFeedback") long idFeedback, FeedbackGym feedbackGym) throws SQLException {
-        FeedbackGymService feedbackGymService = new FeedbackGymServiceImpl(urlDB, userDB, pswDB);
-        feedbackGym.setId(idFeedback);
-        feedbackGymService.updateFeedbackGym(feedbackGym);
+    public Response updateFeedbackGym(@Context SecurityContext securityContext, @PathParam("idFeedback") long idFeedback, FeedbackGym feedbackGym) throws SQLException {
+        if (securityContext.isUserInRole("utente")) {
 
-        return Response.noContent().build();
+            FeedbackGymService feedbackGymService = new FeedbackGymServiceImpl(urlDB, userDB, pswDB);
+            feedbackGym.setId(idFeedback);
+            feedbackGymService.updateFeedbackGym(feedbackGym);
+
+            return Response.noContent().build();
+        }else {
+            return Response.serverError().entity("Non hai i permessi per fare questa operazione").build();
+
+        }
     }
 
     @DELETE
+    @Auth
     @Path("{idFeedback: [0-9]+}")
-    public Response deleteFeedbackGym(@PathParam("idFeedback") long idFeedback) throws SQLException {
-        FeedbackGymService feedbackGymService = new FeedbackGymServiceImpl(urlDB, userDB, pswDB);
-        feedbackGymService.deleteFeedbackGym(idFeedback);
+    public Response deleteFeedbackGym(@Context SecurityContext securityContext,@PathParam("idFeedback") long idFeedback) throws SQLException {
+        if (securityContext.isUserInRole("utente")) {
 
-        return Response.noContent().build();
+            FeedbackGymService feedbackGymService = new FeedbackGymServiceImpl(urlDB, userDB, pswDB);
+            feedbackGymService.deleteFeedbackGym(idFeedback);
+
+            return Response.noContent().build();
+        }else {
+            return Response.serverError().entity("Non hai i permessi per fare questa operazione").build();
+
+        }
+
     }
 
 }
