+++ src/main/java/it/univaq/disim/GymREST/resources/GymRes.java
@@ -1,5 +1,6 @@
 package it.univaq.disim.GymREST.resources;
 
+import javax.annotation.security.RolesAllowed;
 import javax.ws.rs.*;
 import javax.ws.rs.core.*;
 
@@ -57,7 +58,7 @@ public class GymRes {
             User user = userService.getUserByUsername(username);
 
             GymService gymService = new GymServiceImpl(urlDB, userDB, pswDB);
-            gym.setUser(user);
+            gym.setUser(user.getId());
 
             long idGym = gymService.createGym(gym);
 
@@ -69,20 +70,31 @@ public class GymRes {
 
     @PUT
     @Path("{idGym: [0-9]+}")
+    @Auth
     @Consumes(MediaType.APPLICATION_JSON)
-    public Response updateGym(@PathParam("idGym") long idGym, Gym gym) throws SQLException {
-        GymService gymService = new GymServiceImpl(urlDB, userDB, pswDB);
-        gym.setId(idGym);
-        gymService.updateGym(gym);
-        return Response.noContent().build();
+    public Response updateGym(@Context SecurityContext securityContext, @PathParam("idGym") long idGym, Gym gym) throws SQLException {
+        if (securityContext.isUserInRole("gestore")) {
+            GymService gymService = new GymServiceImpl(urlDB, userDB, pswDB);
+            gym.setId(idGym);
+            gymService.updateGym(gym);
+            return Response.noContent().build();
+        } else {
+            return Response.serverError().entity("Non hai i permessi per fare questa operazione").build();
+        }
     }
 
     @DELETE
+    @Auth
     @Path("{idGym: [0-9]+}")
-    public Response deleteGym(@PathParam("idGym") long idGym) throws SQLException {
-        GymService gymService = new GymServiceImpl(urlDB, userDB, pswDB);
-        gymService.deleteGym(idGym);
-        return Response.noContent().build();
+    public Response deleteGym(@Context SecurityContext securityContext, @PathParam("idGym") long idGym) throws SQLException {
+        if (securityContext.isUserInRole("gestore")) {
+            GymService gymService = new GymServiceImpl(urlDB, userDB, pswDB);
+            gymService.deleteGym(idGym);
+            return Response.noContent().build();
+        } else {
+            return Response.serverError().entity("Non hai i permessi per fare questa operazione").build();
+
+        }
     }
